removed auto startup feat, add in pin on load, dismiss the tip in settings -> save to storage, add validation for paths and title fallback used in commands. update CSS to show errors, change license to me.
FYI: removed the automatic startup feature as these tabs remain generally unless a workspace loader is used, in which case they can run the command which will have files loaded. It would also open the sidebar briefly on load which was a bit annyoing.

keeping as v1 as this is unreleased so far